requirements.txt
# Core Agno framework
agno

# OpenAI integration (for embeddings and models)
openai

# Web framework for AgentOS
fastapi[standard]
uvicorn

# Database and vector operations
sqlalchemy
pgvector
psycopg2-binary
psycopg
chonkie

# MCP (Model Context Protocol)
mcp

# Search tools
ddgs
exa-py
googlesearch-python

# Additional utilities
python-dotenv
firecrawl-py
youtube-transcript-api
pycountry
anthropic
slack-sdk
pypdf
yfinance
google-genai
newspaper4k
lxml_html_clean
chonkie
# sentence-transformers  # Optional: for semantic embeddings in knowledge base

# Required for various agent tools
pydantic
rich
starlette
